feat(user): implement auth guard for updating and removing the user

+++ packages/backend/src/user/user.controller.ts
@@ -1,20 +1,45 @@
-import { Body, Controller, Delete, Param, Patch } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Param,
+  Patch,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from '../auth/auth.guard';
+import { GetUser } from './user.decorator';
+import { User } from './entities/user.entity';
 
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  // @todo: Allow only the authenticated user to update its content
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @UseGuards(AuthGuard)
+  update(
+    @Param('id') id: string,
+    @GetUser() user: User,
+    @Body() updateUserDto: UpdateUserDto,
+  ) {
+    if (user.id !== id) {
+      throw new UnauthorizedException(
+        "You're not authorized to make updates to this account",
+      );
+    }
     return this.userService.update(id, updateUserDto);
   }
 
-  // @todo: Allow only the authenticated user to delete its entity
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard)
+  remove(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== id) {
+      throw new UnauthorizedException(
+        "You're not authorized to remove this user",
+      );
+    }
     return this.userService.remove(id);
   }
 }
